Rename test project and modules folder
