+++ tests/KRPC.MechJeb.Test.pyproj
@@ -1,0 +1,93 @@
+<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <SchemaVersion>2.0</SchemaVersion>
+    <ProjectGuid>6fa795a1-14ec-4cf8-9dc7-c828f6b0980a</ProjectGuid>
+    <ProjectHome>
+    </ProjectHome>
+    <StartupFile>TestRunner.py</StartupFile>
+    <SearchPath>
+    </SearchPath>
+    <WorkingDirectory>.</WorkingDirectory>
+    <OutputPath>.</OutputPath>
+    <Name>KRPC.MechJeb.Test</Name>
+    <RootNamespace>tests</RootNamespace>
+    <IsWindowsApplication>False</IsWindowsApplication>
+    <SuppressConfigureTestFrameworkPrompt>true</SuppressConfigureTestFrameworkPrompt>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DebugSymbols>true</DebugSymbols>
+    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <DebugSymbols>true</DebugSymbols>
+    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="modules\ManeuverPlanner.py" />
+    <Compile Include="modules\Maneuver\TimeSelector.py" />
+    <Compile Include="modules\Maneuver\OperationResonantOrbit.py" />
+    <Compile Include="modules\Maneuver\OperationSemiMajor.py" />
+    <Compile Include="modules\Maneuver\OperationTransfer.py" />
+    <Compile Include="modules\Maneuver\OperationLambert.py" />
+    <Compile Include="modules\Maneuver\OperationLan.py" />
+    <Compile Include="modules\Maneuver\OperationLongitude.py" />
+    <Compile Include="modules\Maneuver\OperationMoonReturn.py" />
+    <Compile Include="modules\Maneuver\OperationPlane.py" />
+    <Compile Include="modules\Maneuver\OperationPeriapsis.py" />
+    <Compile Include="modules\Maneuver\OperationKillRelVel.py" />
+    <Compile Include="modules\Maneuver\OperationInterplanetaryTransfer.py" />
+    <Compile Include="modules\Maneuver\OperationInclination.py" />
+    <Compile Include="modules\Maneuver\OperationEllipticize.py" />
+    <Compile Include="modules\Maneuver\OperationCourseCorrection.py" />
+    <Compile Include="modules\Maneuver\OperationCircularize.py" />
+    <Compile Include="modules\Maneuver\OperationApoapsis.py" />
+    <Compile Include="modules\Maneuver\Operation.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="modules\Maneuver\__init__.py" />
+    <Compile Include="modules\ThrustController.py" />
+    <Compile Include="modules\TargetController.py" />
+    <Compile Include="modules\StagingController.py" />
+    <Compile Include="modules\RcsController.py" />
+    <Compile Include="modules\NodeExecutor.py" />
+    <Compile Include="modules\AscentGt.py" />
+    <Compile Include="modules\AscentClassic.py" />
+    <Compile Include="modules\Translatron.py" />
+    <Compile Include="modules\SmartRcs.py" />
+    <Compile Include="modules\SmartAss.py" />
+    <Compile Include="modules\RendezvousAutopilot.py" />
+    <Compile Include="modules\LandingAutopilot.py" />
+    <Compile Include="modules\Annotations.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="modules\AscentAutopilot.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="modules\ComputerModule.py" />
+    <Compile Include="modules\AirplaneAutopilot.py" />
+    <Compile Include="modules\DockingAutopilot.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="TestRunner.py" />
+    <Compile Include="modules\TestCase.py">
+      <SubType>Code</SubType>
+    </Compile>
+    <Compile Include="modules\__init__.py">
+      <SubType>Code</SubType>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="modules\" />
+    <Folder Include="modules\Maneuver\" />
+  </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
+  <!-- Uncomment the CoreCompile target to enable the Build command in
+       Visual Studio and specify your pre- and post-build commands in
+       the BeforeBuild and AfterBuild targets below. -->
+  <!--<Target Name="CoreCompile" />-->
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+</Project>
