--- tests/krpc-mechjeb-tests.pyproj
@@ -1,93 +1,0 @@
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>6fa795a1-14ec-4cf8-9dc7-c828f6b0980a</ProjectGuid>
-    <ProjectHome>
-    </ProjectHome>
-    <StartupFile>TestRunner.py</StartupFile>
-    <SearchPath>
-    </SearchPath>
-    <WorkingDirectory>.</WorkingDirectory>
-    <OutputPath>.</OutputPath>
-    <Name>krpc-mechjeb-tests</Name>
-    <RootNamespace>tests</RootNamespace>
-    <IsWindowsApplication>False</IsWindowsApplication>
-    <SuppressConfigureTestFrameworkPrompt>true</SuppressConfigureTestFrameworkPrompt>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <DebugSymbols>true</DebugSymbols>
-    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <DebugSymbols>true</DebugSymbols>
-    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="krpcmjtest\ManeuverPlanner.py" />
-    <Compile Include="krpcmjtest\Maneuver\TimeSelector.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationResonantOrbit.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationSemiMajor.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationTransfer.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationLambert.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationLan.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationLongitude.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationMoonReturn.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationPlane.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationPeriapsis.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationKillRelVel.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationInterplanetaryTransfer.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationInclination.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationEllipticize.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationCourseCorrection.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationCircularize.py" />
-    <Compile Include="krpcmjtest\Maneuver\OperationApoapsis.py" />
-    <Compile Include="krpcmjtest\Maneuver\Operation.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="krpcmjtest\Maneuver\__init__.py" />
-    <Compile Include="krpcmjtest\ThrustController.py" />
-    <Compile Include="krpcmjtest\TargetController.py" />
-    <Compile Include="krpcmjtest\StagingController.py" />
-    <Compile Include="krpcmjtest\RcsController.py" />
-    <Compile Include="krpcmjtest\NodeExecutor.py" />
-    <Compile Include="krpcmjtest\AscentGt.py" />
-    <Compile Include="krpcmjtest\AscentClassic.py" />
-    <Compile Include="krpcmjtest\Translatron.py" />
-    <Compile Include="krpcmjtest\SmartRcs.py" />
-    <Compile Include="krpcmjtest\SmartAss.py" />
-    <Compile Include="krpcmjtest\RendezvousAutopilot.py" />
-    <Compile Include="krpcmjtest\LandingAutopilot.py" />
-    <Compile Include="krpcmjtest\Annotations.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="krpcmjtest\AscentAutopilot.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="krpcmjtest\ComputerModule.py" />
-    <Compile Include="krpcmjtest\AirplaneAutopilot.py" />
-    <Compile Include="krpcmjtest\DockingAutopilot.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="TestRunner.py" />
-    <Compile Include="krpcmjtest\TestCase.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="krpcmjtest\__init__.py">
-      <SubType>Code</SubType>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="krpcmjtest\" />
-    <Folder Include="krpcmjtest\Maneuver\" />
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
-  <!-- Uncomment the CoreCompile target to enable the Build command in
-       Visual Studio and specify your pre- and post-build commands in
-       the BeforeBuild and AfterBuild targets below. -->
-  <!--<Target Name="CoreCompile" />-->
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-</Project>
