+++ tests/modules/Maneuver/OperationCircularize.py
