+++ tests/modules/Maneuver/OperationEllipticize.py
