+++ tests/modules/SmartAss.py
