+++ tests/modules/StagingController.py
